requirements.txt
flask>=3.0.0
anthropic>=0.40.0
twilio>=9.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
